Fonts work!! well need to adjust titles with the font weight.
